requirements.txt
discord.py 
yt_dlp 
ffmpeg-python
dotenv